Improve wordings
