--- documentation/commands/completion.md
@@ -1,11 +1,0 @@
-<h1 textrun="command-heading">Completion command</h1>
-
-<blockquote textrun="command-summary">
-Generate completion script
-</blockquote>
-
-#### Usage
-
-<pre textrun="command-usage">
-git town completion [bash|zsh|fish|powershell]
-</pre>
